Move permissions helper function to RBAC
