+++ src/ggrc/fulltext/mysql.py
@@ -22,6 +22,7 @@ from ggrc.models import all_models
 from ggrc.models.inflector import get_model
 from ggrc.utils import query_helpers
 from ggrc.rbac import context_query_filter
+from ggrc.rbac import permissions
 from ggrc.fulltext.sql import SqlIndexer
 
 
@@ -70,7 +71,7 @@ class MysqlIndexer(SqlIndexer):
     """
     type_queries = []
     for model_name in model_names:
-      contexts, resources = query_helpers.get_context_resource(
+      contexts, resources = permissions.get_context_resource(
           model_name=model_name,
           permission_type=permission_type,
           permission_model=permission_model
