+++ src/ggrc/rbac/permissions.py
@@ -144,3 +144,24 @@ def has_conditions(action, resource):
   return bool(_permissions.get(action, {})
               .get(resource, {})
               .get('conditions', {}))
+
+
+def get_context_resource(model_name, permission_type='read',
+                         permission_model=None):
+  """Get allowed contexts and resources."""
+  permissions_map = {
+      "create": (create_contexts_for, create_resources_for),
+      "read": (read_contexts_for, read_resources_for),
+      "update": (update_contexts_for, update_resources_for),
+      "delete": (delete_contexts_for, delete_resources_for),
+  }
+
+  contexts = permissions_map[permission_type][0](
+      permission_model or model_name)
+  resources = permissions_map[permission_type][1](
+      permission_model or model_name)
+
+  if permission_model and contexts:
+    contexts = set(contexts) & set(read_contexts_for(model_name))
+
+  return contexts, resources
